+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Unreleased](https://github.com/gilzoide/lua-gdextension/compare/0.2.0...HEAD)
 ### Added
 - Editor plugin that registers the Lua REPL tab, where you can try Lua code using an empty `LuaState`
+- Support for calling Godot String methods using Lua strings
 
 ### Changed
 - The GDExtension is now marked as reloadable